Implement SafeArgs to pass data between Fragments.
This is an addition to the previous commit "Implement Navigation Component".
Thanks to the usage of SafeArgs, we can now directly pass object to our FruitDetailFragment.
It makes the code easier to read and safer!

This change requires to update our Entities Data Classes to make them Parcelable.
It should have be done since their implementation but better late than never, I guess...

Also, second big update with this commit: We now target JVM 1.8 to benefict from the lateset kotlin features,
including `by navArgs()` for SafeArgs!