Implementation of ViewModel
With this commit, we now have an implementation of ViewModel to manage the fruit list in our fragments.
We use a Shared ViewModel design so the VM is shared by both of our fragments, FruitListFragment and FruitDetailFragment.
A shared VM design implies the usage of a nested navigation graph.

With this commit, we also removed all business logic from the FruitListAdapter. It now triggers an itemClickListener lambda
so the fragment is now fully responsible of both, navigation and data management through the VM.
Also, FruitDetailFragment don't need anymore an hard reference to the displayed fruit because it's now handled by the VM.

This commit is also an big step forward for the application as we no more need to keep the legacy DummyContent.
Also, the project ui package will probably need to be restructured as the Shared ViewModel clearly shows I think it wrong.