Add Retrofit to fetch data from remote server.
This commit adds Retrofit dependency to manage network API.
To ease the translation from the remote API response to our
Entities, I changed Nutrition.kt to better mimic the JSON we receive.

For now, data fetching is done under the responsibility of the ViewModel.
This is a big concern regarding the separation of concerns principle and the
scope of our data which is now bound to the VM lifecycle.

Other notable changes:
  * Compile Options updated to target JAVA 1.8 ;
  * Add "android.permission.INTERNET" permission in manifest ;
  * Add "usesCleartextTraffic" to allow unencrypted HTTP traffic.